Make joystick work!

+++ drivers/input_analog_joystick/input_analog_joystick.h
@@ -9,7 +9,7 @@ struct joystick_data {
   const struct device *adc;
   const struct device *dev;
   struct adc_sequence adc_seq;
-  int32_t buffer[2];
+  int16_t buffer[2];
   struct k_timer timer;
   struct k_work work;
 };
